Major object model refactoring.
* Moved the namespace definitions and manager into their own module.
* Alias the namespace in the class as “ns”.
* Separated out the WebAnnotation classes into their own module, and they no longer inherit from pcdm.
* Moved some common core code from pcdm.Resource to ldp.Resource.
* Cleaned up unused imports.
